before collision check

+++ Assets/Game/Scripts/Logic/Authoring/Components/HealthAuthoringComponent.cs
@@ -5,6 +5,6 @@ namespace Game.Authoring
     [Serializable]
     public struct HealthAuthoringComponent
     {
-        public float Health;
+        public int Health;
     }
 }
